topi_ == 0 isn't the same as "topi is zero"
See 4.8.1.2 in AIA: "When topi is zero, this creates a spurious external
interrupt for the hart." I interpret this as meaning "when reading topi
would return zero", which can happen not only when topi_ is 0, but also
when topi_ is >= (a non-0) threshold. In this case, the model was
failing to trigger spurious interrupts.