Remove claimi_ member
We don't need an additional claimi_ member since claimi it has the same
value as topi. Simply deleting it breaks code that relies on sizeof(Idc)
being exactly 32 (as required by the spec), but rather than adding
another reserved field to make this still work, I replace all the uses
of sizeof(Idc) with 32. This seems less error-prone to me.