interruptCount now means total number of sources
Previously, interruptCount was one greater than the actual number of
sources for the APLIC. For example, if the APLIC had 31 sources,
interruptCount would be 32. I think this is confusing, so I changed the
definition.

Note that the meaning is now equivalent to "N" in the AIA spec, such as
in section 4.1 where it says "Each of an APLIC's interrupt sources has a
fixed unique identity number in the range 1 to N, where N is the total
number of sources at the APLIC".