Fix masks for target CSRs
The target mask actually depends on the current delivery mode for a
domain (e.g. when in direct delivery mode, there is no "guest index"
field), which means that the target masks either need to be changed
whenever the delivery mode changes (which would require iterating over
all the domains and updating the masks for targets of active sources) or
the mask needs to be decided at the time the target is written to. I've
chosen to implement the latter option in this commit.

Note that there is some unfortunate complexity here in that the mask
serves two different purposes: one purpose is to decide which specific
bits are writable, which is usually a constant but can be dynamic
in cases such as this where the layout of a CSR depends on the value in other
CSR(s). The other purpose of the mask is is to control whether the CSR
is currently writable at all or must be read-only zero.

Whether or not a given target CSR is read-only zero depends on the value
of the corresponding sourcecfg in the current domain, as well as all the
parent domains going back to the root domain. Instead of inspecting all
of these CSRs at the time of a write to a target CSR, we cache the
writability of the target CSR in the mask. Whenever a sourcecfg changes,
the relevant target masks are recalculated.

So, this means that we have one mask (the one which decides if a target
CSR is currently read-only zero) being cached and one mask (the one
which decides which bits are writable given the domain's current
delivery mode) being determined at the time of writing to the target.
This is error-prone and should be changed.

Because most configurations will probably have very few domains and
interrupt sources and because CSRs will probably be written to
infrequently, the best option is to recalculate everything as necessary
instead of trying to cache things.