Fixing writes to sourcecfg
This fixes several bugs related to writing sourcecfg CSRs. The spec
requires that:

- For an interrupt domain below the root, interrupt sources not
  delegated down to that domain appear to the domain as being not
  implemented
- When source i is not implemented, or appears in some domain not to be
  implemented, sourcecfg[i] is read-only zero in that domain
- Whenever interrupt source i is inactive in an interrupt domain, the
  corresponding pending and enable bits within the domain are read-only
  zeros, and register target[i] is also read-only zero

Note the implications:

- When a source becomes inactive in a domain, the target CSR and its
  mask must be set to 0 (this was not being done)
- When a source which was being delegated to a child domain is no longer
  delegated to that domain after the write to the sourcecfg, the child
  domain must set its corresponding sourcecfg and its mask to 0; this
  must happen recursively (this was not being done)
- If the source was being delegated to a child domain, the child index
  must be captured before the write to the sourcecfg so that we know
  which child domain to update (this was not being done).
- Its possible to simply change which child domain a source is delegated
  to without changing whether it is delegated. This still requires
  deactivating the source in the old child domain (as well as making it
  active in the new child domain).

There are some remaining things to handle. Specifically, the spec says
that "Any write to a sourcecfg register might (or might not) cause the
corresponding interrupt-pending bit to be set to one if the rectified
input value is high (= 1) under the new source mode". This means that we
need to reexamine the source states and determine if any interrupts
should becoming pending under the new source mode.

There are probably also bugs with in_clrip. The spec says that "A read
of an in_clrip register returns the rectified input values of the
corresponding interrupt sources". The rectified input values are cached
in the Inclrip registers when a source state changes, but this cache can
be invalidated by a write to sourcecfg.