Add implementation parameters to DomainParams
Several fields have been added to the DomainParams struct:
- ipriolen: the number of implemented bits for the iprio field of target
- eiidlen: the number of implemented bits for the eiid field of target
  and genmsi
- direct_mode_supported: indicates support for direct delivery mode
- msi_mode_supported: indicates support for msi delivery mode
- le_supported: indicates support for little-endian
- be_supported: indicates support for big-endian

These parameters affect how CSRs such as domaincfg and target are
legalized.

The hart_index field of target is now allowed to be any 14-bit value and
hart_indices can be left empty if only MSI delivery mode is expected to
be used. hart_indices is consulted only when updating topi and when
delivering interrupts via direct delivery mode (only included harts can
be sent interrupts in direct delivery mode).

DomainParams has been moved to Domain.hpp so that the Domain constructor
can use it. This cuts down on the number of places that need to be
updated when adding new parameters as may happen in the future.